Testing Map Connecting
Created Route 50 as a test route and connected it to the left side of
the starting town. It connected perfectly, but there's some black border
issues that need to be addressed.

The game uses a pseudo open world approach where all the maps are
combined into a larger map over a black background, so if your map is
too small sometimes the camera can see that background. This can be
fixed by changing the map size, or adding additional "Fluff" maps that
are just trees, or out of bounds scenery. (The benefit is we can re-use
these!)