More Map Connecting
Tested out adding a fluff map of trees to the south side of Route 50, it
worked. It looks funky because now it doesn't line up with Route 8's
trees on part of the bottom but it's easily fixed (and on our actual
maps we would build around any camera discrepancies like that), more a
proof of concept than anything else.