Finalized Lava Surf Code (Still one issue)
Lava Surfing is now in game and mostly functional after much headache
and the fabled "Eureka" moment.

Problem was the function to check for lava surfing ending was never
being called, which was allowing the user to surf on whatever terrain
was available and surfable.

The issue we now have, is we are unable to jump the player onto
impassable terrain, so attempting to get on to the "lava" tiles fails.
It's probably a simple fix I'm just missing, but it shall be figured
out!