Lava Surf Update
Didn't make any progress towards fixing the passability issues, but
learned that it has nothing to do with the tileset. Something about the
scripting is off. Regular surf is doing something that Lava Surf isn't
that allows it to operate on impassable tiles, I just haven't figure out
what that is yet.